requirements.txt
torch
transformers
safetensors
accelerate
bitsandbytes
python-pptx
PyPDF2
gradio
langchain
langchain-community
ollama